+++ danger-kotlin-sample-plugin/src/main/kotlin/systems/danger/samples/plugin/SamplePlugin.kt
@@ -3,10 +3,10 @@ package systems.danger.samples.plugin
 import systems.danger.kotlin.sdk.DangerPlugin
 
 object SamplePlugin : DangerPlugin() {
-    override val id: String
-        get() = "systems.danger.kotlin.samplePlugin"
+  override val id: String
+    get() = "systems.danger.kotlin.samplePlugin"
 
-    fun myCustomCheck() {
-        context.message("✅ Custom plugin successfully linked")
-    }
+  fun myCustomCheck() {
+    context.message("✅ Custom plugin successfully linked")
+  }
 }